Removed disrespecting pardoy text and emoji
Removing a trans flag emoji which was added to previous commit just as a parody and I don't want to hurt any community and pretend to be someone in order to just have fun. Removing something which does not represent my identity also respect to those people who really identify as such. I am sorry for pretending. Thank you and take love everyone! <3